Refactored project, worked on judgment prediction datasetset creation
