+++ scrc/annotation/.gitignore
@@ -1,16 +1,10 @@
 prodigy_results.db
 .env
-<<<<<<< HEAD
-documentation/
 wheel/
-judgment_explainability/datasets/*.jsonl
-judgment_explainability/datasets/dataset_scrc/
-judgment_explainability/datasets/sjp/
-=======
-documentation
-wheel
-judgment_explainability/datasets/*.jsonl
-judgment_explainability/datasets/dataset_scrc
-judgment_explainability/datasets/sjp
->>>>>>> 8ac356fcd1a84c36276e94e51b66104b7940d745
-annotations/
+judgment_explainability/datasets/
+judgment_explainability/documentation
+judgment_explainability/annotations/
+judgment_prediction/datasets
+prodigy_dataset_creation/dataset_scrc
+prodigy_dataset_creation/sjp
+
